Create a new 'helpers' module for 'documentation'
